Refactor phase 3f: split the maintenance routes into a package
webapp/admin/maintenance.py (1,802 lines) becomes the
webapp/admin/maintenance/ package: 15 modules plus a 118-line __init__,
every one within the 400-line guidance. All 31 top-level definitions are
ast.dump()-identical.

  blueprint, paths, operations, eas_settings   leaves
  noaa                  -> its own constants
  serialization         -> its own helper
  routes_operations     -> blueprint, operations, paths
  routes_database       -> blueprint
  routes_env            -> blueprint, paths
  routes_poll           -> blueprint
  routes_location       -> blueprint
  routes_import         -> blueprint, noaa
  routes_alerts         -> blueprint, noaa, serialization
  routes_expiry         -> blueprint
  routes_eas_settings   -> blueprint, eas_settings

The __file__ hazard fired for the second phase running, with a wider
blast radius than certbot's. repo_root resolves tools/create_backup.py
and tools/inplace_upgrade.py, is passed as their cwd, and locates the
.env that Settings -> Environment reads and writes. One directory short
and all three point into webapp/: the backup invokes a script that does
not exist and the env editor edits a file nothing reads. Fixed to four
hops and pinned.

__all__ is not the public surface. get_operation_status is a route
handler, absent from the module's ten-name __all__, and imported by
app_core/websocket_push.py inside a function whose caller swallows
failures — the same shape as the AudioSourceConfigDB regression. The
export test AST-walks the tree for every
`from webapp.admin.maintenance import ...` instead of listing names.

tests/test_maintenance_package.py is this module's first coverage: 20
tests. Two pin things that are easy to get wrong later — maintenance_bp
takes no url_prefix and writes /admin into each decorator, the opposite
of the certbot blueprint next door; and _OPERATION_STATE is mutated in
place, never rebound, which is what makes a by-value import of it safe.
Both hazards mutation-checked.

app.py was assessed and deliberately not started. It is not a package
split: it builds the app at import time and configures it with 103
module-level statements, so statement order is the behaviour and neither
AST equality nor a URL-map diff can catch a reorder. It needs the
singleton turned into a real factory first, as its own commit. Recorded
as 3g in the plan with the steps it actually requires.

Verification: 31/31 AST matches, URL map unchanged at 549 rules / 0
differences, repo_root resolves to its pre-split value, full suite 2287
passed.
