+++ docs/reference/protocols/MDC1200.md
@@ -453,7 +453,7 @@ to wake every subscriber on a quiet channel.
 These were added by Alembic migration
 [`20260505_add_mdc1200_to_eas_settings`](../../../app_core/migrations/versions/20260505_add_mdc1200_to_eas_settings.py).
 The auto-heal `_PENDING_MIGRATIONS` block in
-[`webapp/admin/maintenance.py`](../../../webapp/admin/maintenance.py) mirrors
+[`webapp/admin/maintenance/eas_settings.py`](../../../webapp/admin/maintenance/eas_settings.py) mirrors
 the same DDL so installs that update DB-first also self-heal.
 
 ### 5.2 Configuration plumbing
@@ -491,7 +491,7 @@ OTA-relayed, and manual broadcasts.
 Settings → Pre/Post-Alert Signaling** with the dropdown options gated by
 `data-eas-tone-show="mdc1200"` and `mdc1200-custom`. The `custom`
 sub-fields accept either decimal (0..255) or `0x..` hex notation;
-`webapp/admin/maintenance.py` parses both via `int(value, 0)`.
+`webapp/admin/maintenance/routes_eas_settings.py` parses both via `int(value, 0)`.
 
 ---
